jenkins file updated to publish reports to github pages

+++ Jenkinsfile
@@ -24,10 +24,7 @@ pipeline {
         stage('Build and Generate Reports') {
             steps {
                 echo 'Building and generating reports...'
-                script {
-                    def mavenCommand = isUnix() ? 'mvn -Dmaven.test.failure.ignore clean test -Dcucumber.filter.tags=@regression' : 'mvn -Dmaven.test.failure.ignore clean test -Dcucumber.filter.tags=@regression'
-                    sh mavenCommand
-                }
+                sh 'mvn -Dmaven.test.failure.ignore clean test -Dcucumber.filter.tags=@regression'
             }
         }
 
@@ -35,10 +32,13 @@ pipeline {
             steps {
                 echo 'Publishing Allure results to GitHub Pages...'
                 script {
-                    def removeCommand = isUnix() ? 'rm -rf allure-report' : 'rmdir /s /q allure-report'
-                    def moveCommand = isUnix() ? 'mv allure-report/* docs/' : 'move allure-report\\* docs\\'
+                    def gitInit = !fileExists('.git')
+                    if (gitInit) {
+                        sh 'git init'
+                        sh 'git remote add origin https://github.com/mahmoud2911/piplineDemo.git'
+                    }
 
-                    sh removeCommand  // Remove existing allure-report directory
+                    sh 'rm -rf allure-report' // Remove existing allure-report directory
                     allure([
                         includeProperties: true,
                         jdk: '',
@@ -46,22 +46,22 @@ pipeline {
                             [path: 'allure-results']
                         ]
                     ])
-                    sh 'mkdir -p docs'  // Create a docs directory to host Allure reports on GitHub Pages
-                    sh moveCommand  // Move Allure reports to the docs directory
+                    sh 'mkdir -p docs' // Create a docs directory to host Allure reports on GitHub Pages
+                    sh 'mv allure-report/* docs/' // Move Allure reports to the docs directory
 
-                    // Configure Git
-                    sh 'git init'
-                    sh 'git remote add origin https://github.com/mahmoud2911/piplineDemo.git'
+                    if (gitInit) {
+                        sh 'git add .'
+                        sh 'git commit -m "Initial commit"'
+                        sh 'git push -u origin master'
+                    } else {
+                        sh 'git rm -r docs/*'
+                        sh 'git commit -m "Remove old Allure reports"'
+                        sh 'git push -u origin master'
 
-                    // Delete old reports from GitHub Pages
-                    sh 'git rm -r docs/*'
-                    sh 'git commit -m "Remove old Allure reports"'
-                    sh 'git push -u origin master'
-
-                    // Add and commit the latest reports
-                    sh 'git add docs/*'
-                    sh 'git commit -m "Add latest Allure reports"'
-                    sh 'git push -u origin master'
+                        sh 'git add docs/*'
+                        sh 'git commit -m "Add latest Allure reports"'
+                        sh 'git push -u origin master'
+                    }
                 }
             }
         }
